requirements.txt
# Core LangGraph dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Common AI/ML dependencies
openai>=1.0.0
anthropic>=0.34.0
tiktoken>=0.7.0

# Azure OpenAI integration
langchain-openai>=0.2.0
azure-identity>=1.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web framework (optional)
fastapi>=0.104.0
uvicorn>=0.24.0

# Development tools
jupyter>=1.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0